refactor(v1alpha1): split casting into envelope and per-kind files
casting.go now contains only the envelope: Casting struct, JSONSchemaOneOf,
UnmarshalJSON dispatch, TrackableProperties dispatch.

casting_signoz.go contains SigNoz-specific types (SigNozCastingSpec,
SigNozCastingStatus, kindSigNozCasting variant), accessors (SigNozSpec,
SigNozStatus), factories (DefaultSigNozCasting, ExampleSigNozCasting), and
the unmarshal/trackable helpers.

Dropped kind-agnostic DefaultCasting/ExampleCasting wrappers in favor of
explicit per-kind names; updated call sites in gen.go and yamlconfig.

Tests split per file: casting_test.go for envelope behavior,
casting_signoz_test.go for SigNoz factories and accessor invariants.

Adding a new kind = one new casting_<kind>.go file plus a switch arm in
each dispatcher.
