+++ cmd/foundryctl/gen.go
@@ -66,7 +66,7 @@ func runGenExamples(ctx context.Context, logger *slog.Logger) error {
 	for deployment := range foundry.Registry.CastingItems() {
 		logger.InfoContext(ctx, "generating example files for deployment", slog.Any("deployment", deployment))
 
-		config := v1alpha1.ExampleCasting()
+		config := v1alpha1.ExampleSigNozCasting()
 		config.SigNozSpec().Deployment = deployment
 
 		rootPath := filepath.Join("docs", "examples/", deployment.Platform.String(), deployment.Mode.String(), deployment.Flavor.String())
