+++ internal/config/yamlconfig/config.go
@@ -37,7 +37,7 @@ func (config *yamlConfig) GetV1Alpha1(ctx context.Context, path string) (v1alpha
 		return v1alpha1.Casting{}, fmt.Errorf("failed to unmarshal yaml: %w", err)
 	}
 
-	defaultCasting := v1alpha1.DefaultCasting()
+	defaultCasting := v1alpha1.DefaultSigNozCasting()
 	// merge overrides into defaults (base)
 	if err := v1alpha1.MergeCasting(&defaultCasting, &casting); err != nil {
 		return v1alpha1.Casting{}, fmt.Errorf("failed to merge default casting: %w", err)
